Añadidos los permisos para el crud de Comedor

+++ proyectoHostal/apps/views.py
@@ -80,40 +80,45 @@ def ComedorListar(request):
     return render(request, 'Cocina/menu_listar.html', {'menus' : menus}) 
 
 def ComedorAgregar(request):
-    if request.method == 'POST':
-        form = MenuForms(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()            
-            tipo_menu = form.cleaned_data.get('tipo_menu')
-            messages.success(request, f'El Menu {tipo_menu} Se Ha Agregado!')
-            return redirect('/comedor/listar')
-    else:
-        form = MenuForms()
-    return render(request, 'Cocina/menu_agregar.html', {'form': form}) 
+    if request.user.groups.filter(name = "GERENTE COCINA" ).exists() or request.user.groups.filter(name = "ADMINISTRADOR" ).exists() or request.user.is_superuser:
+        if request.method == 'POST':
+            form = MenuForms(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()            
+                tipo_menu = form.cleaned_data.get('tipo_menu')
+                messages.success(request, f'El Menu {tipo_menu} Se Ha Agregado!')
+                return redirect('/comedor/listar')
+        else:
+            form = MenuForms()
+        return render(request, 'Cocina/menu_agregar.html', {'form': form})
+    return redirect('/comedor/listar')
 
 def ComedorEditar(request,menu_id):
-    instancia= Menu.objects.get(id=menu_id)
-    if instancia.documento_menu:
-        path = instancia.documento_menu.url
-    else:
-        path = None
-        
-    form=  MenuForms(instance=instancia)
-    if request.method=="POST":
-        form= MenuForms(request.POST, request.FILES, instance=instancia)
-        if form.is_valid():
-            instancia= form.save(commit=False)
-            instancia.save()
-            tipo_menu = form.cleaned_data.get('tipo_menu')
-            messages.success(request, f'El Menu {tipo_menu} Se Ha Agregado!')
-            return redirect('/comedor/listar')
-    return render(request, 'Cocina/menu_editar.html', {'form': form,'path':path}) 
+    if request.user.groups.filter(name = "GERENTE COCINA" ).exists() or request.user.groups.filter(name = "ADMINISTRADOR" ).exists() or request.user.is_superuser:
+        instancia= Menu.objects.get(id=menu_id)
+        if instancia.documento_menu:
+            path = instancia.documento_menu.url
+        else:
+            path = None
+        form=  MenuForms(instance=instancia)
+        if request.method=="POST":
+            form= MenuForms(request.POST, request.FILES, instance=instancia)
+            if form.is_valid():
+                instancia= form.save(commit=False)
+                instancia.save()
+                tipo_menu = form.cleaned_data.get('tipo_menu')
+                messages.success(request, f'El Menu {tipo_menu} Se Ha Agregado!')
+                return redirect('/comedor/listar')
+        return render(request, 'Cocina/menu_editar.html', {'form': form,'path':path})
+    return redirect('/comedor/listar')
 
 def ComedorEliminar(request,menu_id):
-    instacia= Menu.objects.get(id=menu_id)
-    instacia.delete()
-    messages.warning(request, f'El Menu {instacia.tipo_menu} Se Ha Eliminado!')
-    return redirect('/comedor/listar') 
+    if request.user.groups.filter(name = "GERENTE COCINA" ).exists() or request.user.groups.filter(name = "ADMINISTRADOR" ).exists() or request.user.is_superuser:
+        instacia= Menu.objects.get(id=menu_id)
+        instacia.delete()
+        messages.warning(request, f'El Menu {instacia.tipo_menu} Se Ha Eliminado!')
+        return redirect('/comedor/listar') 
+    return redirect('/comedor/listar')
 
 def ComedorAdjunto(request,menu_id):
     instacia= Menu.objects.get(id=menu_id)
